fix(osx, android sdk): search ~/Applications and /Applications
OSX can install applications system wide, or for certain users. And what is chosen is not always known up front.

On my system, I had the reverse of what was expected, my sdk was installed in ~, and when I installed the Jetbrains Toolbox, I installed it into /.

The current source, expects the opposite, without a good reason or explanation of why this is preferred.

This PR allows the locations of both the tool box and the android sdk to exist in either.

Additionally, this PR addresses that Node 20.9 at least, will not resolve an escaped space in `path.join`, and instead fail, as it is escaping the directory fragments.

However, the command execution DOES require escaping of that part.

Test-Plan:

1. Run envinfo previously:
```
➤ npx envinfo --ides

  IDEs:
    IntelliJ: 2023.3.5
    Vim: 9.0 - /usr/bin/vim
    Xcode: 15.3/15E204a - /usr/bin/xcodebuild

```
2. Apply fixes

3. run envinfo
```
➤ npm run start

> envinfo@7.11.1 start
> node -r esm src/cli.js
...
  IDEs:
    Android Studio: 2023.2 AI-232.10300.40.2321.11567975
    IntelliJ: 2023.3.5
    Vim: 9.0 - /usr/bin/vim
    Xcode: 15.3/15E204a - /usr/bin/xcodebuild
...
```

4. Run tests
```
➤ npm run test

> envinfo@7.11.1 test
> jest --env=node && eslint src && prettier -l src/**/*.js

 PASS  __tests__/scopedPackages.test.js
 PASS  __tests__/duplicates.test.js
 PASS  __tests__/envinfo.test.js
 PASS  __tests__/utils.test.js

Test Suites: 4 passed, 4 total
Tests:       51 passed, 51 total
Snapshots:   13 passed, 13 total
Time:        0.449s, estimated 1s
Ran all test suites.
```